First commit -MSVC project file -Engine side modifications/hacks

+++ source/duke3d/src/sounds_mapster32.cpp
@@ -56,7 +56,7 @@ int32_t g_numEnvSoundsPlaying;
 
 void MUSIC_Update(void) {}  // needed when linking
 
-void S_Callback(uint32_t);
+void S_Callback(intptr_t);
 
 /*
 ===================
@@ -440,7 +440,7 @@ void S_Update(void)
         }
 }
 
-void S_Callback(uint32_t num)
+void S_Callback(intptr_t num)
 {
     int32_t i,j,k;
